requirements.txt
# --- Framework e Server API ---
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.6.4
pydantic_core==2.16.3

# --- Database ---
pymongo[srv]==4.6.1
influxdb-client==1.41.0
psycopg2-binary==2.9.10

# --- Model Inference ---
tensorflow==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-estimator==2.15.0
keras==2.15.0
flatbuffers==25.2.10
numpy==1.26.4
ml-dtypes==0.2.0

# --- Test e Analisi ---
requests==2.31.0
pytest==8.1.1
pytest-asyncio==0.23.6
pandas==2.2.3
matplotlib==3.10.1
seaborn==0.13.2

# --- Locust ---
locust==2.35.0
gevent==24.11.1
greenlet==3.1.1
geventhttpclient==2.3.3

# --- OAuth (per TensorFlow/Google API) ---
google-auth==2.38.0
google-auth-oauthlib==1.2.1
oauthlib==3.2.2
rsa==4.9
pyasn1==0.6.1
pyasn1_modules==0.4.1

# --- Utility ---
python-dotenv==1.1.0
typing_extensions==4.12.2
watchfiles==1.0.4
httpx==0.27.0
httpcore==1.0.7
h11==0.14.0
sniffio==1.3.1
click==8.1.8
colorama==0.4.6
tqdm==4.67.1

# --- Logging e TensorBoard ---
Markdown==3.7
tensorboard==2.15.2
tensorboard-data-server==0.7.2
Werkzeug==3.1.3
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
certifi==2025.1.31
